chore: split embedders in individual files
