+++ projects/pgai/pgai/vectorizer/embedders/__init__.py
@@ -1,0 +1,5 @@
+from .ollama import Ollama
+from .openai import OpenAI
+from .voyageai import VoyageAI
+
+__all__ = ["Ollama", "OpenAI", "VoyageAI"]
